+++ vkconfig_gui/tab_applications.cpp
@@ -41,15 +41,6 @@ static void PathInvalid(const Path &path, const char *message) {
 
 TabApplications::TabApplications(MainWindow &window, std::shared_ptr<Ui::MainWindow> ui)
     : Tab(TAB_APPLICATIONS, window, ui), _launch_application(nullptr) {
-    this->ui->launch_executable_search->setIcon(::Get(::ICON_FILE_SEARCH));
-    this->ui->launch_executable_append->setIcon(::Get(::ICON_FILE_APPEND));
-    this->ui->launch_executable_remove->setIcon(::Get(::ICON_FILE_REMOVE));
-    this->ui->launch_options_append->setIcon(::Get(::ICON_OPTIONS_COPY));
-    this->ui->launch_options_remove->setIcon(::Get(::ICON_OPTIONS_REMOVE));
-    this->ui->launch_options_dir_button->setIcon(::Get(::ICON_FOLDER_SEARCH));
-    this->ui->launch_options_log_button->setIcon(::Get(::ICON_FILE_SEARCH));
-    this->ui->launch_options_log_open->setIcon(::Get(::ICON_FILE_EXPORT));
-
     this->connect(this->ui->launch_executable_list, SIGNAL(currentIndexChanged(int)), this,
                   SLOT(on_launch_executable_list_activated(int)));
     this->connect(this->ui->launch_executable_list->lineEdit(), SIGNAL(textEdited(QString)), this,